+++ R/field_types.R
@@ -197,10 +197,13 @@ NULL
 #'   field level, using [readr::col_datetime()] format specifications, e.g.
 #'   `format = "%d/%m/%Y"`. When a format is supplied, it must match the
 #'   complete string.
+#' @param na Column-specific vector of strings that should be interpreted as missing
+#'   values (in addition to those specified at dataset level)
 #' @rdname field_types_available
 #' @export
 ft_timepoint <- function(includes_time = TRUE,
-                         format = "") {
+                         format = "",
+                         na = NULL) {
   # NOTE: nonconformant values appear in validation warnings
   agg_fun <- c("n")
   options <- NULL
@@ -214,6 +217,7 @@ ft_timepoint <- function(includes_time = TRUE,
     collector = readr::col_datetime(format = format),
     data_class = "POSIXct",
     aggregation_functions = agg_fun,
+    na = na,
     options = options
   )
 }
@@ -221,9 +225,11 @@ ft_timepoint <- function(includes_time = TRUE,
 #' @section Details: `ft_uniqueidentifier()` - identifies data fields which
 #'   contain a (usually computer-generated) identifier for an entity, e.g. a
 #'   patient. It does not need to be unique within the dataset.
+#' @param na Column-specific vector of strings that should be interpreted as missing
+#'   values (in addition to those specified at dataset level)
 #' @rdname field_types_available
 #' @export
-ft_uniqueidentifier <- function() {
+ft_uniqueidentifier <- function(na = NULL) {
   # TODO: potential additional aggregation_functions:
   # proportion numeric; distinct first chars
   field_type(
@@ -237,7 +243,8 @@ ft_uniqueidentifier <- function() {
       "min_length",
       "max_length",
       "mean_length"
-    )
+    ),
+    na = na
   )
 }
 
@@ -247,9 +254,12 @@ ft_uniqueidentifier <- function() {
 #'   generated for each distinct subcategory as well as for the field overall.
 #'   If `FALSE`, aggregated values will only be generated for the field overall.
 #'   Default = `FALSE`
+#' @param na Column-specific vector of strings that should be interpreted as missing
+#'   values (in addition to those specified at dataset level)
 #' @rdname field_types_available
 #' @export
-ft_categorical <- function(aggregate_by_each_category = FALSE) {
+ft_categorical <- function(aggregate_by_each_category = FALSE,
+                           na = NULL) {
   # TODO: allow more options for aggregate_by_each_category,
   # e.g. topx (bysize), or accept a vector of values
   agg_fun <- c("n", "missing_n", "missing_perc", "distinct")
@@ -264,6 +274,7 @@ ft_categorical <- function(aggregate_by_each_category = FALSE) {
     collector = readr::col_character(),
     data_class = "character",
     aggregation_functions = agg_fun,
+    na = na,
     options = options
   )
 }
@@ -272,9 +283,11 @@ ft_categorical <- function(aggregate_by_each_category = FALSE) {
 #' `ft_numeric()` - identifies data fields which contain numeric values that
 #' should be treated as continuous. Any values which contain non-numeric
 #' characters (including grouping marks) will be classed as non-conformant
+#' @param na Column-specific vector of strings that should be interpreted as missing
+#'   values (in addition to those specified at dataset level)
 #' @rdname field_types_available
 #' @export
-ft_numeric <- function() {
+ft_numeric <- function(na = NULL) {
   field_type(
     type = "numeric",
     collector = readr::col_double(),
@@ -289,7 +302,8 @@ ft_numeric <- function() {
       "max",
       "mean",
       "median"
-    )
+    ),
+    na = na
   )
 }
 
@@ -304,10 +318,13 @@ ft_numeric <- function() {
 #'   field level, using [readr::col_datetime()] format specifications, e.g.
 #'   `format = "%d/%m/%Y"`. When a format is supplied, it must match the
 #'   complete string.
+#' @param na Column-specific vector of strings that should be interpreted as missing
+#'   values (in addition to those specified at dataset level)
 #' @rdname field_types_available
 #' @export
 ft_datetime <- function(includes_time = TRUE,
-                        format = "") {
+                        format = "",
+                        na = NULL) {
   agg_fun <-
     c(
       "n",
@@ -329,34 +346,41 @@ ft_datetime <- function(includes_time = TRUE,
     collector = readr::col_datetime(format = format),
     data_class = "POSIXct",
     aggregation_functions = agg_fun,
+    na = na,
     options = options
   )
 }
 
 #' @section Details: `ft_freetext()` - identifies data fields which contain
 #'   free text values. Only presence/missingness will be evaluated.
+#' @param na Column-specific vector of strings that should be interpreted as missing
+#'   values (in addition to those specified at dataset level)
 #' @rdname field_types_available
 #' @export
-ft_freetext <- function() {
+ft_freetext <- function(na = NULL) {
   field_type(
     type = "freetext",
     collector = readr::col_character(),
     data_class = "character",
-    aggregation_functions = c("n", "missing_n", "missing_perc")
+    aggregation_functions = c("n", "missing_n", "missing_perc"),
+    na = na
   )
 }
 
 #' @section Details: `ft_simple()` - identifies data fields where you only
 #'   want presence/missingness to be evaluated (but which are not necessarily
 #'   free text).
+#' @param na Column-specific vector of strings that should be interpreted as missing
+#'   values (in addition to those specified at dataset level)
 #' @rdname field_types_available
 #' @export
-ft_simple <- function() {
+ft_simple <- function(na = NULL) {
   field_type(
     type = "simple",
     collector = readr::col_character(),
     data_class = "character",
-    aggregation_functions = c("n", "missing_n", "missing_perc")
+    aggregation_functions = c("n", "missing_n", "missing_perc"),
+    na = na
   )
 }
 
@@ -434,6 +458,7 @@ field_type <- function(type,
                        collector,
                        data_class,
                        aggregation_functions = c("n", "missing_n", "missing_perc"),
+                       na = NULL,
                        options = NULL) {
   structure(
     list(
@@ -441,6 +466,7 @@ field_type <- function(type,
       collector = collector,
       data_class = data_class,
       aggregation_functions = aggregation_functions,
+      na = na,
       options = options
     ),
     class = c(paste0("daiquiri_field_type_", type), "daiquiri_field_type")
@@ -538,7 +564,6 @@ field_type_aggregation_functions <- function(field_type) {
 }
 
 
-
 # -----------------------------------------------------------------------------
 #' Convert field_types object to string for printing
 #'
@@ -553,6 +578,12 @@ field_types_to_string <- function(field_types) {
       names(field_types[ft]), "\t",
       "<", field_type_type(field_types[[ft]]), ">"
     )
+    if (!is.null(field_types[[ft]]$na)) {
+      s <- paste0(
+        s, "\t",
+        "na: ", paste(dQuote(field_types[[ft]]$na, q = FALSE), collapse = ", ")
+      )
+    }
     if (!is.null(field_types[[ft]]$options)) {
       s <- paste0(
         s, "\t",